+++ packages/@uppy/aws-s3/tests/s3-client/minio.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, inject, it, vi } from 'vitest'
 import { S3mini } from '../../src/s3-client/S3.js'
 import { createSigV4Signer } from '../../src/s3-client/signer.js'
 import { randomBytes } from '../test-utils/browser-crypto.js'
-import { beforeRun, cleanupTestBeforeAll } from './_shared.test.js'
+import { beforeRun, cleanupTestBeforeAll } from './test-util.js'
 import type { BucketConfigs } from './teardown.js'
 
 const name = 'minio'
